refactor: move ProviderTestFixture to test_utils
