+++ crates/edr_optimism/src/hardfork.rs
@@ -126,7 +126,7 @@ fn chain_configs() -> &'static HashMap<u64, &'static ChainConfig<OptimismChainSp
     CONFIGS.get_or_init(|| {
         let mut hardforks = HashMap::new();
         hardforks.insert(10, mainnet_config());
-        hardforks.insert(11_155_111, sepolia_config());
+        hardforks.insert(11_155_420, sepolia_config());
 
         hardforks
     })
